fixed a warning
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@21633 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/terminal/Coding.cpp
@@ -51,7 +51,7 @@ get_nth_encoding(int i, int *id)
 	if (id == NULL)
 		return B_BAD_VALUE;
 	
-	if (i < 0 || i >= sizeof(kEncodingTable) / sizeof(etable))
+	if (i < 0 || i >= (int)(sizeof(kEncodingTable) / sizeof(etable)))
 		return B_BAD_INDEX;
 
 	*id = kEncodingTable[i].id;
